+++ layout/columns1.php
@@ -28,43 +28,4 @@
 
 defined('MOODLE_INTERNAL') || die;
 
-// Include header.
-$sidepostdrawer = false;
-require_once(dirname(__FILE__) . '/includes/header.php');
-// Include secondary navigation.
-require_once(dirname(__FILE__) . '/includes/secondarynav.php');
-?>
-
-<div id="maincontainer" class="container outercont">
-    <?php
-        echo $OUTPUT->get_news_ticker();
-        echo $OUTPUT->page_navbar();
-    ?>
-    <div id="page-content" class="row">
-        <div id="region-main-box" class="col-12">
-            <section id="region-main">
-            <?php
-            echo $OUTPUT->course_content_header();
-            if (!empty($secondarynavigation)) {
-                echo $secondarynavigation;
-            }
-            echo $OUTPUT->activity_header();
-            if (!empty($overflow)) {
-                echo $overflow;
-            }
-            echo $OUTPUT->main_content();
-            echo $OUTPUT->activity_navigation();
-            echo $OUTPUT->course_content_footer();
-            ?>
-            </section>
-        </div>
-    </div>
-</div>
-
-<?php
-if (empty($PAGE->layout_options['nofooter'])) {
-    // Include footer.
-    require_once(dirname(__FILE__) . '/includes/footer.php');
-} else {
-    require_once(dirname(__FILE__) . '/includes/nofooter.php');
-}
+echo $OUTPUT->columns_one_layout();